Conflicts:
	settings/2S_SEH.xml
	settings/2S_SEH_crate.xml

+++ HWInterface/CommandProcessorInterface.h
@@ -1,10 +1,10 @@
 #ifndef _CommandProcessorInterface_H__
 #define _CommandProcessorInterface_H__
 
-#include "../Utils/Utilities.h"
-#include "../Utils/easylogging++.h"
-#include "BeBoard.h"
-#include "RegManager.h"
+#include "HWDescription/BeBoard.h"
+#include "HWInterface/RegManager.h"
+#include "Utils/Utilities.h"
+#include "Utils/easylogging++.h"
 #include <string>
 
 namespace Ph2_HwInterface
@@ -24,7 +24,7 @@ class CommandProcessorInterface : public RegManager
         LOG(ERROR) << BOLDRED << __PRETTY_FUNCTION__ << "\tError: implementation of virtual member function CommandProcessorInterface::WriteCommand is absent" << RESET;
     }
 
-    virtual std::vector<uint32_t> ReadReply(uint8_t pNWords)
+    virtual std::vector<uint32_t> ReadReply(int pNWords)
     {
         LOG(ERROR) << BOLDRED << __PRETTY_FUNCTION__ << "\tError: implementation of virtual member function CommandProcessorInterface::ReadReply is absent" << RESET;
         return {};
